requirements.txt
django==5.0.7
djangorestframework==3.15.2
django-filter==24.2
djangorestframework-jsonapi[django-filter]==7.0.2
mysqlclient==2.2.4